+++ examples/style-guide-new/components/colors/index.styl
@@ -1,0 +1,165 @@
+@import '../../../../build/stylesheets/defaults/index.css'
+@import '../../../../build/stylesheets/mixins/common.styl'
+
+@keyframes fadeContentIn
+  from
+    transform translateY(30px)
+    opacity 0
+  to
+    opacity 0.9
+
+section
+  margin-bottom 50px
+  padding-bottom 50px
+  border-bottom 1px solid grey-100
+
+.colors
+  width 70vw
+  animation fadeContentIn 500ms forwards
+
+.color-title
+  color #4C6072
+  font-size 24px
+
+.color-description
+  padding 16px 0px 45px 0px
+
+.color-chips
+  display flex
+  flex-direction row
+  flex-wrap wrap
+
+  .swatch
+    background-color #FFFFFF
+    border 1px solid #E5EAEF
+    border-radius 6px
+    display flex
+    flex-direction column
+    height 250px
+    margin 0 30px 45px 0
+    width 200px
+    
+    .swatch-color
+      border-radius 6px 6px 0px 0px
+      width 100%
+      height 180px
+    
+    .swatch-label
+      border-top 1px solid #E5EAEF
+      flex-grow 1
+      
+      .color-name
+        display block
+        color #4C6072
+        font-size 18px
+        margin 12px 0 0 15px
+      
+      .color-hex
+        color #9CACBB
+        font-size 16px
+        margin 3px 0 0 15px
+
+.code-snippet
+  left -60px
+  padding-left 15px
+  position relative
+
+  .call-out
+    color grey-900
+    display inline-block
+    text-transform uppercase
+    &.extra
+      margin-top 30px
+    & > strong
+      margin-left 15px
+    label
+      text-transform none
+    p
+      line-height 4px
+      margin-left 45px
+      text-transform none
+
+  .snippet-container
+    background-color grey-50
+    border 1px solid grey-200
+    border-radius 6px
+    -moz-border-radius 6px
+    -webkit-border-radius 6px
+    font-size 14px
+    margin 30px 0 0 45px
+    position relative
+
+  .snippet-header
+    color grey-700
+    border-bottom 1px solid grey-200
+    border-radius 6px 6px 0px 0px
+    -moz-border-radius 6px 6px 0px 0px
+    -webkit-border-radius 6px 6px 0px 0px
+    padding 15px
+    label
+      text-transform uppercase
+      letter-spacing 1.2px
+
+  .code-field
+    background-color white
+    border-left 1px solid grey-200
+    border-radius 0px 0px 6px 0px
+    -moz-border-radius 0px 0px 6px 0px
+    -webkit-border-radius 0px 0px 6px 0px
+    font-family monospace
+    margin 0 0 0 45px
+    padding 15px
+    pre
+      color grey-900
+      display inline-block
+      letter-spacing 1.5px
+      line-height 20px
+      margin -2px 0 0 -45px
+      white-space pre-wrap
+      &:nth-child(2n)
+        margin-left 45px
+
+  pre
+    span
+      color mp-red
+    b
+      color grey-900
+      font-weight 400
+
+.color-library
+  display flex
+  flex-direction row
+  flex-wrap wrap
+  
+  .color-group-container
+    display flex
+    flex-direction column
+    height 400px
+    margin-left 5px
+    width 30vw
+  
+    .color-group-title
+      color #9CACBB
+      padding 50px 0 10px 0
+
+    .color-group
+      display flex
+      flex-direction column
+      flex-grow 1
+
+      .color-segment
+        display flex
+        flex-direction row
+        flex-grow 1
+        justify-content space-between
+        width 100%
+
+        .color-segment-hex
+          color white
+          font-size 18px
+          opacity 0.5
+          padding 20px
+
+        .color-segment-name
+          color white
+          padding 20px
